Create a map of peeks and calculate diff
We have a map of peeks, where some of them are unknown (then we get only the names of those) and some of them are known (then we have the height). 

Our job is to store the map and calculate the difference between the highest peek and the lowest one.